file-manager-8

+++ Nodejs-Project/file-manager/src/app/controller/ManagerController.ts
@@ -1,11 +1,17 @@
 import { Request, Response, NextFunction } from "express";
-import folderBuilder from "../../helper/fbuilder";
-import folderRemove from "../../helper/fremove";
+/* import folderBuilder from "../../helper/fbuilder";
+import folderRemove from "../../helper/fremove"; */
 import docBuilder from "../../helper/dbuilder";
 import fs from "fs"
 import path  from "path"
+import docRemover from "../../helper/dremover";
 const { PrismaClient } = require('@prisma/client');
 const prisma = new PrismaClient();
+const rootPath = process.cwd();
+
+function uniqueId(len:number):number{
+   return Math.floor(Math.random() * len)
+}
 
 export async function view(req:Request,res:Response){
       const categories = await prisma.category.findMany({
@@ -15,14 +21,13 @@ export async function view(req:Request,res:Response){
                   }
             }},
       })
-    
+      
       return res.render('index',{categories:categories}) 
 }
 
 export async function folder(req:Request,res:Response){
       const categories: string = req.query.categories as string;
       const name: string = req.query.name as string; 
-      // const type: string = req.query.type as string; 
       const build = await prisma.file.create({
             data: {
                   categories: {
@@ -31,12 +36,7 @@ export async function folder(req:Request,res:Response){
                   folder_name:name
             },
       })
-    //  folderBuilder(categories,name)
-
-    
-      
-      
-      
+   
       return res.redirect('back')
 }
 export async function privateFolder(req:Request,res:Response){
@@ -56,64 +56,60 @@ export async function privateFolder(req:Request,res:Response){
             })
             //folderBuilder(categories,name)
       }else if(type=='docs'){
-            const fileName = name+'.txt'
-            const build = await prisma.docs.create({
-                  data: {
-                        files: {
-                              connect: { id: +parent } // Assuming id is the primary key of the Category table
+            const fileName = name+"_"+uniqueId(1000000000)+'.txt'
+            if(docBuilder(fileName)){
+                  const build = await prisma.docs.create({
+                        data: {
+                              files: {
+                                    connect: { id: +parent } // Assuming id is the primary key of the Category table
+                              },
+                              type:'txt',
+                              name:fileName
                         },
-                        type:'txt',
-                        name:fileName
-                  },
-            })
-            docBuilder(fileName)
+                  })
+            }
+            
+            
            
       }
-      
-      
-      
       return res.redirect('back')
 }
 
 export async function remove(req:Request,res:Response) {
       if(req.params.id){
-           
-           /*  const file = await prisma.file.findUnique({
-                  where: {
-                      id: +(req.params.id),
-                  },
-            }); */
-           
-            /* if(folderRemove(file.categoriesId+"/"+file.parent,file.folder_name)){
-                 
-            } */
-             
             const deleteFile = await prisma.file.delete({
                   where: {
                     id: +(req.params.id),
                   },
-            })
-            return res.redirect('back')
+            })   
       }
+      return res.redirect('back')
 }
 
 export async function docremove(req:Request,res:Response) {
       if(req.params.id){
-            console.log(req.params.id)
-            const file = await prisma.docs.delete({
+            const docs = await prisma.docs.findUnique({
                   where: {
                       id: +(req.params.id),
                   },
             });
-       
+            
+            if(docRemover(docs.name)){
+                  const file = await prisma.docs.delete({
+                        where: {
+                        id: +(req.params.id),
+                        },
+                  }); 
+            }
+            
             return res.redirect('back')
       }
 }
 
 export async function viewFile(req:Request,res:Response){
      
-      const dirname = path.resolve(__filename,'../../../../'); // Use __filename instead of __dirname
-      const folderPath = path.join(dirname, "storage");
+     // const dirname = path.resolve(__filename,'../../../../'); // Use __filename instead of __dirname
+      const folderPath = path.join(rootPath, "storage");
       const doc = await prisma.docs.findUnique({
             where: {
                 id: +(req.params.id),
@@ -129,7 +125,7 @@ export async function viewFile(req:Request,res:Response){
 export async function saveFile(req:Request,res:Response){
      
       const dirname = path.resolve(__filename,'../../../../'); // Use __filename instead of __dirname
-      const folderPath = path.join(dirname, "storage");
+      const folderPath = path.join(rootPath, "storage");
       const doc = await prisma.docs.findUnique({
             where: {
                 id: +(req.params.id),
@@ -147,6 +143,13 @@ export async function saveFile(req:Request,res:Response){
             status:true
       })
      
+}
+
+export async function truncate(req:Request,res:Response){
+      const table = req.params.table
+      await prisma[table].deleteMany();
+      res.send('Table Truncated')
+      res.end()
 }
